add css to TweetSearch and reorganize files
