+++ frontend/components/TweetSearch/tweetSearch.css
@@ -1,0 +1,13 @@
+.app-container {
+    width: 50%;
+    margin: 0 auto;
+}
+
+input {
+    width: 100%;
+    line-height: 2;
+    margin-top: 15px;
+    text-align: center;
+    font-size: 1.3em;
+    border: 1px solid black;
+}
